+++ server/src/main/java/org/opensearch/common/network/NetworkModule.java
@@ -55,6 +55,7 @@ import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.http.HttpServerTransport;
 import org.opensearch.index.shard.PrimaryReplicaSyncer.ResyncTask;
 import org.opensearch.plugins.NetworkPlugin;
+import org.opensearch.plugins.SecureTransportSettingsProvider;
 import org.opensearch.ratelimitting.admissioncontrol.enums.AdmissionControlActionType;
 import org.opensearch.tasks.RawTaskStatus;
 import org.opensearch.tasks.Task;
@@ -67,6 +68,7 @@ import org.opensearch.transport.TransportRequestHandler;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -85,6 +87,9 @@ public final class NetworkModule {
     public static final String HTTP_TYPE_KEY = "http.type";
     public static final String HTTP_TYPE_DEFAULT_KEY = "http.type.default";
     public static final String TRANSPORT_TYPE_DEFAULT_KEY = "transport.type.default";
+    public static final String TRANSPORT_SSL_ENFORCE_HOSTNAME_VERIFICATION_KEY = "transport.ssl.enforce_hostname_verification";
+    public static final String TRANSPORT_SSL_ENFORCE_HOSTNAME_VERIFICATION_RESOLVE_HOST_NAME_KEY = "transport.ssl.resolve_hostname";
+    public static final String TRANSPORT_SSL_DUAL_MODE_ENABLED_KEY = "transport.ssl.dual_mode.enabled";
 
     public static final Setting<String> TRANSPORT_DEFAULT_TYPE_SETTING = Setting.simpleString(
         TRANSPORT_TYPE_DEFAULT_KEY,
@@ -93,6 +98,22 @@ public final class NetworkModule {
     public static final Setting<String> HTTP_DEFAULT_TYPE_SETTING = Setting.simpleString(HTTP_TYPE_DEFAULT_KEY, Property.NodeScope);
     public static final Setting<String> HTTP_TYPE_SETTING = Setting.simpleString(HTTP_TYPE_KEY, Property.NodeScope);
     public static final Setting<String> TRANSPORT_TYPE_SETTING = Setting.simpleString(TRANSPORT_TYPE_KEY, Property.NodeScope);
+
+    public static final Setting<Boolean> TRANSPORT_SSL_ENFORCE_HOSTNAME_VERIFICATION = Setting.boolSetting(
+        TRANSPORT_SSL_ENFORCE_HOSTNAME_VERIFICATION_KEY,
+        true,
+        Property.NodeScope
+    );
+    public static final Setting<Boolean> TRANSPORT_SSL_ENFORCE_HOSTNAME_VERIFICATION_RESOLVE_HOST_NAME = Setting.boolSetting(
+        TRANSPORT_SSL_ENFORCE_HOSTNAME_VERIFICATION_RESOLVE_HOST_NAME_KEY,
+        true,
+        Property.NodeScope
+    );
+    public static final Setting<Boolean> TRANSPORT_SSL_DUAL_MODE_ENABLED = Setting.boolSetting(
+        TRANSPORT_SSL_DUAL_MODE_ENABLED_KEY,
+        false,
+        Property.NodeScope
+    );
 
     private final Settings settings;
 
@@ -151,9 +172,17 @@ public final class NetworkModule {
         HttpServerTransport.Dispatcher dispatcher,
         ClusterSettings clusterSettings,
         Tracer tracer,
-        List<TransportInterceptor> transportInterceptors
+        List<TransportInterceptor> transportInterceptors,
+        Collection<SecureTransportSettingsProvider> secureTransportSettingsProvider
     ) {
         this.settings = settings;
+
+        if (secureTransportSettingsProvider.size() > 1) {
+            throw new IllegalArgumentException(
+                "there is more than one secure transport settings provider: " + secureTransportSettingsProvider
+            );
+        }
+
         for (NetworkPlugin plugin : plugins) {
             Map<String, Supplier<HttpServerTransport>> httpTransportFactory = plugin.getHttpTransports(
                 settings,
@@ -170,6 +199,7 @@ public final class NetworkModule {
             for (Map.Entry<String, Supplier<HttpServerTransport>> entry : httpTransportFactory.entrySet()) {
                 registerHttpTransport(entry.getKey(), entry.getValue());
             }
+
             Map<String, Supplier<Transport>> transportFactory = plugin.getTransports(
                 settings,
                 threadPool,
@@ -182,6 +212,43 @@ public final class NetworkModule {
             for (Map.Entry<String, Supplier<Transport>> entry : transportFactory.entrySet()) {
                 registerTransport(entry.getKey(), entry.getValue());
             }
+
+            // Register any secure transports if available
+            if (secureTransportSettingsProvider.isEmpty() == false) {
+                final SecureTransportSettingsProvider secureSettingProvider = secureTransportSettingsProvider.iterator().next();
+
+                final Map<String, Supplier<HttpServerTransport>> secureHttpTransportFactory = plugin.getSecureHttpTransports(
+                    settings,
+                    threadPool,
+                    bigArrays,
+                    pageCacheRecycler,
+                    circuitBreakerService,
+                    xContentRegistry,
+                    networkService,
+                    dispatcher,
+                    clusterSettings,
+                    secureSettingProvider,
+                    tracer
+                );
+                for (Map.Entry<String, Supplier<HttpServerTransport>> entry : secureHttpTransportFactory.entrySet()) {
+                    registerHttpTransport(entry.getKey(), entry.getValue());
+                }
+
+                final Map<String, Supplier<Transport>> secureTransportFactory = plugin.getSecureTransports(
+                    settings,
+                    threadPool,
+                    pageCacheRecycler,
+                    circuitBreakerService,
+                    namedWriteableRegistry,
+                    networkService,
+                    secureSettingProvider,
+                    tracer
+                );
+                for (Map.Entry<String, Supplier<Transport>> entry : secureTransportFactory.entrySet()) {
+                    registerTransport(entry.getKey(), entry.getValue());
+                }
+            }
+
             List<TransportInterceptor> pluginTransportInterceptors = plugin.getTransportInterceptors(
                 namedWriteableRegistry,
                 threadPool.getThreadContext()